+++ tests/bfabric_cli/test_cli_api_delete.py
@@ -5,8 +5,8 @@ from bfabric_scripts.cli.api.delete import Params, cmd_api_delete
 
 @pytest.fixture
 def mock_client(mocker):
-    client = mocker.Mock(spec=Bfabric)
-    client.config.base_url = "http://test-bfabric.com"
+    client = mocker.MagicMock(spec=Bfabric)
+    client.config.base_url = "http://test-bfabric.com/bfabric"
     return client